System/unlock/requirements.txt
aiohttp==3.6.2
asgiref==3.2.7
astroid==2.4.0
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.2
Babel==2.8.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
Django==3.0.6
django-filter==2.2.0
django-fontawesome==1.0
django-phonenumber-field==4.0.0
django-storages==1.9.1
djangorestframework==3.11.0
firebase-admin==4.2.0
gcloud==0.17.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.0
imutils==0.5.3
isort==4.3.21
jws==0.1.3
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
msgpack==1.0.0
multidict==4.7.5
numpy==1.18.4
oauth2client==3.0.0
oauthlib==3.1.0
opencv-python==4.2.0.34
pandas==1.0.3
phonenumbers==8.12.2
Pillow==7.1.2
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.5.2
pytesseract==0.3.4
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
twilio==6.40.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
